requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
yfinance
pandas
configparser
prettytable
python-telegram-bot>=21.9
azure-data-tables
requests
azure-storage-file-share
python-dotenv
opencensus-ext-azure